Updates for live repair tests.
Allow a downstairs to go Faulted->Faulted

Abort a LiveRepair when we get the first error.

If we receive an error during live repair, restart the connection to
the downstairs that had the error, but leave the state as Failed.

Fixed a few of the LiveRepair tests

Moved create_test_upstairs() to just create the upstairs.
Current callers of that should instead use start_up_and_repair(ClientId)